Make changes according to gh comments.
